requirements.txt
anyio==3.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
bcrypt==4.0.0
beautifulsoup4==4.12.2
bleach==6.0.0
cffi==1.15.1
click==8.1.3
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.85.0
fastjsonschema==2.17.1
fqdn==1.5.1
h11==0.14.0
idna==3.4
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
overrides==7.3.1
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.3
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.7
python-multipart==0.0.5
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.20.4
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
uri-template==1.2.0
uvicorn==0.18.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.3
widgetsnbextension==4.0.7
